+++ D-Velop.novaextension/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Version 0.9.6
 
+Format Selection.
+
+Menu access to extension preferences and restart server.
+
 With the release of Nova 10.4, the problem with serialization of 1 and 0
 in the LSP client is gone, and the associated workarounds and limits are
 removed. It is recommended to upgrade to Nova 10.4 or better.